+++ src/main/scala/de/tdng2011/game/kernel/Entity.scala
@@ -5,9 +5,9 @@ import actors.Actor
 
 abstract class Entity(var pos : Vec2, val publicId : Long) extends Actor {
 
-  protected var direction : Double;
-  protected var speed : Int  // m/s
-  protected var radius : Int  // m
+  protected var direction : Float;
+  protected var speed : Short  // m/s
+  protected var radius : Short  // m
   protected val entityType : EntityTypes.Value
 
   var thinkHandler : PartialFunction[Any, Option[EntityDescription]] = { case _ => None }
